used euro2020 plugin for predgame plugin
